Refactored out arff formatting
